% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxview.R
\name{boxview}
\alias{boxview}
\title{Display code logic in nested boxes}
\usage{
boxview(fun, width = 200, optimization = c("weak", "medium", "strong"))
}
\arguments{
\item{fun}{A function}

\item{width}{The desired width, the output will be wider if the we can't
make it narrow enough by wrapping code, and will be narrower if we don't
need the width when not wrapping at all}

\item{optimization}{The level of space optimization, to wrap calls only
when needed. \code{boxview()} is quite slow, and strong optimization can take
a long time for big functions. You might not need higher levels most of
the time, especially with sufficient width where the difference is often
barely noticeable.}
}
\value{
a 'boxview' object that prints the boxed code to the console.
}
\description{
\code{boxview()} displays the code of a function in nested boxes, showing the
'yes' and 'no' clauses of \code{if} calls side by side. It often makes code easier to
skim than the traditional way.
}
\examples{
\dontrun{
boxview(ave)
boxview(ave, width = 10)
boxview(ave, width = 10, optimization = "medium")
boxview(interaction)
boxview(data.frame)
}
}
